Refuse a date unless TRANSFORM_DATES expands it, dropping the plain-number reading
With #1225 refusing a genuine datetime column outright, reading one as
nanoseconds since the epoch has no reason left to exist: `NumericalDateTransformer`,
the resolution scaling it needed, the "hold dates" warning and the two-strategy
factory all go. One concrete `DateTransformer` remains. Off, a point in time is
refused with #1225's error naming it, at fit and at predict alike; on, it is
expanded through skrub as before. A duration still becomes its length in seconds
either way, since that loses nothing and leaving it alone only crashes
validation. A period is now a point in time like any other: refused or expanded,
never read as a number.
