+++ changelog/1227.added.md
@@ -1,1 +1,1 @@
-`inference_config={"TRANSFORM_DATES": True}` expands a column holding a genuine datetime dtype (`datetime64`, tz-aware, or `period`) into calendar features (the year, the day of year, the seconds since the epoch, and the cyclical month, day and weekday pairs, plus the time of day when the column carries one) instead of reading it as one plain number. Off by default.
+`inference_config={"TRANSFORM_DATES": True}` expands a column holding a genuine datetime dtype (`datetime64`, tz-aware, or `period`) into calendar features (the year, the day of year, the seconds since the epoch, and the cyclical month, day and weekday pairs, plus the time of day when the column carries one), where such a column is otherwise refused. Off by default.
