+++ changelog/1227.fixed.1.md
@@ -1,1 +1,1 @@
-Fitting or predicting on a `period` or `timedelta64` column no longer aborts. A period becomes the instant it starts at, and a duration its length in seconds.
+Fitting or predicting on a `timedelta64` column no longer aborts: a duration becomes its length in seconds. A `period` column is handled like a `datetime64` one, refused with a clear error or expanded with `TRANSFORM_DATES`, instead of aborting with `Unknown dtype: period[D]`.
