+++ src/tabpfn/preprocessing/datetimes/columns.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-import numpy as np
 import pandas as pd
 
 if TYPE_CHECKING:
@@ -36,39 +35,11 @@ def as_timestamp(column: pd.Series) -> pd.Series:
     return column
 
 
-def to_nanoseconds(column: pd.Series) -> pd.Series:
-    """Cast one point-in-time column to nanoseconds since the epoch.
-
-    Scaled by the column's own resolution, since `astype("int64")` counts ticks
-    of that resolution rather than nanoseconds: pandas 3 reads a date as
-    `datetime64[us]` where pandas 2 read it as `[ns]`, and the same timestamps
-    have to come out as the same number either way, or a fit and a predict of
-    one column that arrived at different resolutions are scaled apart. Scaled in
-    `float64` rather than by converting the column to `[ns]` first, which raises
-    `OutOfBoundsDatetime` outside 1678-2262.
-
-    As `float64`, so a missing date (`NaT`) survives as `NaN`: `NaT.astype`
-    `("int64")` maps it to that dtype's huge sentinel value instead, which
-    `float64` has no equivalent of. Exact enough for any realistic datetime
-    column: nanosecond-since-epoch magnitudes are still ~256ns short of
-    float64's precision limit for a present-day date, far below anything a
-    tabular feature would need to distinguish.
-    """
-    is_missing = column.isna().to_numpy()
-    # A tz-aware dtype carries `.unit`; a plain one reports it only through numpy.
-    unit = getattr(column.dtype, "unit", None) or np.datetime_data(column.dtype)[0]
-    as_ns = column.astype("int64").astype("float64") * (
-        np.timedelta64(1, unit) / np.timedelta64(1, "ns")
-    )
-    as_ns[is_missing] = np.nan
-    return as_ns
-
-
 def to_seconds(column: pd.Series) -> pd.Series:
     """Cast one `timedelta64` column to its length in seconds.
 
     A duration carries no calendar, so its length is the whole of its meaning:
-    unlike a point in time, there is nothing an expansion could add later.
+    unlike a point in time, there is nothing an expansion could add.
     """
     return column.dt.total_seconds()
