+++ tests/test_preprocessing/test_datetimes/test_conversion.py
@@ -1,59 +1,43 @@
 #  Copyright (c) Prior Labs GmbH 2026.
 
-"""Tests for picking a date transformer and for the predict-time guard."""
+"""Tests for the predict paths' guard around `date_transformer_`."""
 
 from __future__ import annotations
 
 import pandas as pd
+import pytest
 
-from tabpfn.preprocessing.datetimes import (
-    NumericalDateTransformer,
-    SkrubDateTransformer,
-    convert_dates,
-    make_date_transformer,
-)
+from tabpfn.errors import TabPFNUserError
+from tabpfn.preprocessing.datetimes import DateTransformer, convert_dates
 
 
 def _frame(dates: pd.Series | pd.DatetimeIndex | list) -> pd.DataFrame:
     return pd.DataFrame({"num": [1.0, 2.0, 3.0], "date": dates})
 
 
-class TestMakeDateTransformer:
-    """`make_date_transformer`: `TRANSFORM_DATES` picks the implementation."""
-
-    def test__flag_off__reads_dates_as_plain_numbers(self) -> None:
-        transformer = make_date_transformer(transform_dates=False)
-
-        assert isinstance(transformer, NumericalDateTransformer)
-
-    def test__flag_on__expands_dates_into_calendar_features(self) -> None:
-        transformer = make_date_transformer(transform_dates=True)
-
-        assert isinstance(transformer, SkrubDateTransformer)
-
-    def test__declared_categoricals__reach_the_transformer(self) -> None:
-        X = _frame(pd.date_range("2020-01-01", periods=3))
-
-        out = make_date_transformer(
-            categorical_indices=[1], transform_dates=True
-        ).fit_transform(X)
-
-        assert out.X is X
-
-
 class TestConvertDates:
-    """`convert_dates`: the predict paths' guard for an unset attribute."""
+    """`convert_dates`: the predict paths' guard for an unset attribute.
+
+    `fit_from_preprocessed` never sets `date_transformer_`, exactly like the
+    pre-existing `ordinal_encoder_` guard, so the fallback has nothing fitted
+    to expand a date with.
+    """
 
     class _Source:
         def __init__(self, **attributes: object) -> None:
             self.__dict__.update(attributes)
 
-    def test__source_without_a_transformer__still_converts(self) -> None:
-        """`fit_from_preprocessed` never sets `date_transformer_`, exactly like the
-        pre-existing `ordinal_encoder_` guard.
-        """
+    def test__source_without_a_transformer__refuses_a_date(self) -> None:
         X = _frame(pd.date_range("2020-01-01", periods=3))
+
+        with pytest.raises(TabPFNUserError, match=r"1 \('date'\)"):
+            convert_dates(X, self._Source(categorical_features_indices=None))
+
+    def test__source_without_a_transformer__still_converts_a_duration(self) -> None:
+        X = _frame(pd.to_timedelta([1, 2, 3], unit="D"))
+
         out = convert_dates(X, self._Source(categorical_features_indices=None))
+
         assert pd.api.types.is_numeric_dtype(out["date"])
 
     def test__source_without_a_transformer__honours_declared_categoricals(
@@ -66,7 +50,7 @@ class TestConvertDates:
     def test__source_with_a_fitted_transformer__uses_it(self) -> None:
         X = _frame(pd.date_range("2020-01-01", periods=3))
         source = self._Source(
-            date_transformer_=NumericalDateTransformer(categorical_indices=[1]),
+            date_transformer_=DateTransformer(categorical_indices=[1]),
             categorical_features_indices=None,
         )
         assert convert_dates(X, source) is X
